+++ src/features/island/hud/components/settings-menu/general-settings/GeneralSettings.tsx
@@ -12,21 +12,7 @@ export const GeneralSettings: React.FC<ContentComponentProps> = ({
 }) => {
   const { t } = useAppTranslation();
 
-  const {
-    gameService,
-    showAnimations,
-    toggleAnimations,
-    enableQuickSelect,
-    toggleQuickSelect,
-  } = useContext(Context);
-
-  const onToggleAnimations = () => {
-    toggleAnimations();
-  };
-
-  const onToggleQuickSelect = () => {
-    toggleQuickSelect();
-  };
+  const { gameService } = useContext(Context);
 
   return (
     <>
@@ -47,7 +33,6 @@ export const GeneralSettings: React.FC<ContentComponentProps> = ({
           )}
         </Button>
       )}
-
       <Button onClick={() => onSubMenuClick("discord")} className="mb-1">
         <span>{`Discord`}</span>
       </Button>
@@ -58,19 +43,8 @@ export const GeneralSettings: React.FC<ContentComponentProps> = ({
       <Button className="mb-1" onClick={() => onSubMenuClick("appearance")}>
         <span>{t("gameOptions.generalSettings.appearance")}</span>
       </Button>
-      <Button className="mb-1" onClick={onToggleAnimations}>
-        <span>
-          {showAnimations
-            ? t("gameOptions.generalSettings.disableAnimations")
-            : t("gameOptions.generalSettings.enableAnimations")}
-        </span>
-      </Button>
-      <Button className="mb-1" onClick={onToggleQuickSelect}>
-        <span>
-          {enableQuickSelect
-            ? t("gameOptions.generalSettings.disableQuickSelect")
-            : t("gameOptions.generalSettings.enableQuickSelect")}
-        </span>
+      <Button className="mb-1" onClick={() => onSubMenuClick("behaviour")}>
+        <span>{t("gameOptions.generalSettings.behaviour")}</span>
       </Button>
       <Button onClick={() => onSubMenuClick("share")} className="mb-1">
         <span>{t("gameOptions.generalSettings.share")}</span>
